Fix: totalHeading in OTOSLocalizer and PinpointLocalizer would increase in wrong direction with heading numbers that their absolute value wraps past 180°
Thank you to GoldenElf58 from FTC Team 16657 for this fix.

Each update loop, totalHeading would always increase because it was incremented by getSmallestAngleDifference, which is always positive. This multiplies that by the turn direction, so that totalHeading can decrease and work as intended (that being a heading number that wraps past +/-180°)